+++ examples/00-test-script.R
@@ -18,31 +18,36 @@ l2 <- leaflet() %>%
                                             transitMaxTransfers = 0),
                    verbose = FALSE, progress = FALSE)
 
+# test times
+l3 <- leaflet() %>%
+  addProviderTiles("CartoDB.Positron") %>%
+  addTargomoTimes(source_data = data.frame(lat = 51.52, lng = -0.19),
+                  target_data = data.frame(lat = runif(100, 51.47, 51.57),
+                                           lng = runif(100, -0.3, -0.1)),
+                  options = targomoOptions(travelType = "car",
+                                           maxEdgeWeight = "1h"),
+                  drawOptions = timeDrawOptions(palette = "inferno",
+                                                maxTime = 1800,
+                                                fillOpacity = 0.8,
+                                                reverse = TRUE,
+                                                radius = 5,
+                                                weight = 1))
+
+
+
 # manual testing
-options_list <- targomoOptions(travelType = "car")
-source_data <- data.frame(id = c("A", "B"),
-                          lat = c(51.52, 51.53), lng = c(-0.20, -0.19))
+source_data <- data.frame(id = "A", lat = 51.52, lng = -0.19)
 target_data <- data.frame(id = c("X", "Y", "Z"),
-                          lat = c(51.50, NA, 51.51), lng = c(-0.18, -0.185, -0.17))
+                          lat = c(51.50, 51.515, 51.51), lng = c(-0.18, -0.185, -0.17))
 
-options <- deriveOptions(options_list)
-sources <- deriveSources(source_data, options = options, id = ~id)
-targets <- deriveTargets(target_data, id = ~id)
+getTargomoTimes(source_data = source_data, target_data = target_data,
+                source_lat = ~lat, source_lng = ~lng,
+                target_lat = ~lat, target_lng = ~lng,
+                source_id = ~id, target_id = ~id,
+                options = targomoOptions(travelType = c("bike")))
 
-t_body <- createRequestBody("time", sources, targets, options)
+l3 <-
 
-t_response <- callTargomoAPI(service = "time", body = t_body, verbose = TRUE)
-t_payload <- httr::content(t_response)
-
-processTime(t_payload)
-
-
-
-suppressWarnings({
-  getTargomoTimes(source_data = source_data, target_data = target_data,
-                  source_id = ~id, target_id = ~id,
-                  options = targomoOptions(travelType = c("car", "bike")))
-})